+++ oidf-services/src/test/java/se/swedenconnect/oidf/service/suites/InMemoryTestSuite.java
@@ -28,6 +28,7 @@ import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.mock.env.MockEnvironment;
 import se.swedenconnect.oidf.service.Application;
 import se.swedenconnect.oidf.service.CacheTestCases;
+import se.swedenconnect.oidf.service.JWKSTestCases;
 import se.swedenconnect.oidf.service.TestConfiguration;
 import se.swedenconnect.oidf.service.entity.ApplicationReadyEndpoint;
 import se.swedenconnect.oidf.service.entity.RegistryMock;
@@ -57,7 +58,8 @@ import java.util.Random;
     ActuatorTestCases.class,
     ResolverPolicyTestCases.class,
     ResolverCritTestCases.class,
-    ResolverSamlSpTestCases.class
+    ResolverSamlSpTestCases.class,
+    JWKSTestCases.class
 })
 public class InMemoryTestSuite {