+++ oidf-services/src/test/java/se/swedenconnect/oidf/service/suites/RedisTestSuite.java
@@ -30,6 +30,7 @@ import org.springframework.mock.env.MockEnvironment;
 import org.testcontainers.utility.DockerImageName;
 import se.swedenconnect.oidf.service.Application;
 import se.swedenconnect.oidf.service.CacheTestCases;
+import se.swedenconnect.oidf.service.JWKSTestCases;
 import se.swedenconnect.oidf.service.TestConfiguration;
 import se.swedenconnect.oidf.service.entity.ApplicationReadyEndpoint;
 import se.swedenconnect.oidf.service.entity.RegistryMock;
@@ -59,7 +60,8 @@ import java.util.Random;
     ActuatorTestCases.class,
     ResolverPolicyTestCases.class,
     ResolverCritTestCases.class,
-    ResolverSamlSpTestCases.class
+    ResolverSamlSpTestCases.class,
+    JWKSTestCases.class
 })
 public class RedisTestSuite {